+++ launcher/src/com/genir/aitweaks/launcher/hullmods/Skirmisher.kt
@@ -1,0 +1,15 @@
+package com.genir.aitweaks.launcher.hullmods
+
+import com.fs.starfarer.api.combat.BaseHullMod
+import com.fs.starfarer.api.combat.ShipAPI
+
+class Skirmisher : BaseHullMod() {
+    override fun showInRefitScreenModPickerFor(ship: ShipAPI): Boolean {
+        return !ship.isFrigate
+    }
+
+    override fun getDescriptionParam(index: Int, hullSize: ShipAPI.HullSize?): String? = when (index) {
+            0 -> "frigates,"
+            else -> null
+        }
+}